feat(STORY-031): implement hybrid trace collection strategy [GREEN]
Replace fixed-rounds placeholder with adaptive collection loop using:
- Completion signal: stops on status="complete" in A2A response metadata
- Idle detection: stops when agent exceeds idle_threshold_seconds per message
- Timeout safety: hard stop after max_timeout_seconds via asyncio.wait_for

Also widens TraceCollectionConfig.max_timeout_seconds/idle_threshold_seconds
from int to float to support sub-second precision in testing.

+++ src/common/models.py
@@ -120,8 +120,8 @@ class InteractionStep(BaseModel):
 class TraceCollectionConfig(BaseModel):
     """Configurable timeout and idle detection settings for adaptive trace collection."""
 
-    max_timeout_seconds: int = 30
-    idle_threshold_seconds: int = 5
+    max_timeout_seconds: float = 30
+    idle_threshold_seconds: float = 5
     use_completion_signals: bool = True
